# api/v0/gen/swagger.yaml
swagger: "2.0"
info:
  description: "Optikon is an edge computing API for Kubernetes. It exposes Helm and the cluster registry for a comprehensive UI"
  version: "0.0.1"
  title: "Optikon"
  contact:
    email: "stelouie@cisco.com"
basePath: "/v0"
...
schemes:
- "http"
- "https"
paths:
...
  /releases/{releaseId}:
    get:
      tags:
      - "releases"
      summary: "Find release by ID"
      description: "Returns a single release"
      operationId: "getReleaseById"
      produces:
      - "application/json"
      parameters:
      - name: "releaseId"
        in: "path"
        description: "ID of release to return"
        required: true
        type: "string"
      - in: query
        name: labels
        type: string
        description: The node labels to identify applicable clusters
      responses:
        200:
          description: "OK"
          schema:
            type: "object"
            $ref: "#/definitions/release.Release"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
        500:
          $ref: "#/responses/InternalServerError"
    put:
      tags:
      - "releases"
      summary: "Update an existing release"
      description: ""
      operationId: "updateRelease"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "releaseId"
        in: "path"
        description: "ID of release to update"
        required: true
        type: "string"
      - name: chartTar
        in: formData
        type: file
        required: true
        description: The file to upload
      - in: query
        name: labels
        type: string
        description: The node labels to identify applicable clusters
      responses:
        200:
          description: "OK"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
        500:
          $ref: "#/responses/InternalServerError"
    delete:
      tags:
      - "releases"
      summary: "Delete release by ID"
      operationId: "deleteRelease"
      produces:
      - "application/json"
      parameters:
      - name: "releaseId"
        in: "path"
        description: "ID of release to return"
        required: true
        type: "string"
      - in: query
        name: labels
        type: string
        description: The node labels to identify applicable clusters
      responses:
        200:
          description: "OK"
        400:
          $ref: "#/responses/BadRequest"
        401:
          $ref: "#/responses/Unauthorized"
        404:
          $ref: "#/responses/NotFound"
        500:
          $ref: "#/responses/InternalServerError"
definitions:
...
  release.Release:
    type: "object"
    properties:
      Name:
        type: "string"
      Info:
        $ref: "#/definitions/release.Info"
      Chart:
        $ref: "#/definitions/chart.Chart"
      Config:
        $ref: "#/definitions/chart.Config"
      Manifest:
        type: "string"
      Hooks:
        type: "array"
        items:
          $ref: "#/definitions/release.Hook"
      Version:
        type: "integer"
        format: "int32"
      Namespace:
        type: "string"
      OnCluster:
        type: "string"
  release.Info:
    type: "object"
    properties:
      Status:
        $ref: >-
          #/definitions/release.Status
      FirstDeployed:
        type: "string"
      LastDeployed:
        type: "string"
      Deleted:
        type: "string"
      Description:
        type: "string"
  release.Status:
    type: "object"
    properties:
      Code:
        type: "string"
        enum:
          - "UNKNOWN"
          - "DEPLOYED"
          - "DELETED"
          - "SUSPENDED"
          - "FAILED"
          - "DELETING"
          - "PENDING_INSTALL"
          - "PENDING_UPGRADE"
          - "PENDING_ROLLBACK"
      Resources:
        type: "string"
      Notes:
        type: "string"
      LastTestSuiteRun:
        $ref: "#/definitions/release.TestSuite"
  release.Hook:
    type: "object"
    properties:
      Name:
        type: "string"
      Kind:
        type: "string"
      Path:
        type: "string"
      Manifest:
        type: "string"
      Events:
        type: "array"
        items:
          type: "string"
          enum:
            - "UNKNOWN"
            - "PRE_INSTALL"
            - "POST_INSTALL"
            - "PRE_DELETE"
            - "POST_DELETE"
            - "PRE_UPGRADE"
            - "POST_UPGRADE"
            - "PRE_ROLLBACK"
            - "POST_ROLLBACK"
            - "RELEASE_TEST_SUCCESS"
            - "RELEASE_TEST_FAILURE"
      LastRun:
        type: "string"
      Weight:
        type: "integer"
        format: "int32"
      DeletePolicies:
        type: "array"
        items:
          type: "string"
          enum:
            - "SUCCEEDED"
            - "FAILED"
            - "BEFORE_HOOK_CREATION"
  release.TestRun:
    type: "object"
    properties:
      Name:
        type: "string"
      Status:
        type: "string"
      Info:
        type: "string"
      StartedAt:
        type: "string"
      CompletedAt:
        type: "string"
  release.TestSuite:
    type: "object"
    properties:
      StartedAt:
        type: "string"
      CompletedAt:
        type: "string"
      Results:
        type: "array"
        items:
          $ref: "#/definitions/release.TestRun"
  chart.Chart:
    type: "object"
    properties:
      Metadata:
        $ref: "#/definitions/chart.Metadata"
      Template:
        type: "array"
        items:
          $ref: "#/definitions/chart.Template"
      Dependencies:
        type: "array"
        items:
          $ref: "#/definitions/chart.Chart"
      Config:
        $ref: "#/definitions/chart.Config"
      Files:
        type: "array"
        items:
          $ref: "#/definitions/protobuf.Any"
  chart.Metadata:
    type: "object"
    properties:
      Engine:
        type: "string"
        enum:
          - "UNKNOWN"
          - "GOTPL"
      Name:
        type: "string"
      Home:
        type: "string"
      Sources:
        type: "array"
        items:
          type: "string"
      Version:
        type: "string"
      Description:
        type: "string"
      Keywords:
        type: "array"
        items:
          type: "string"
      Maintainers:
        type: "array"
        items:
          $ref: "#/definitions/chart.Maintainer"
      Icon:
        type: "string"
      ApiVersion:
        type: "string"
      Condition:
        type: "string"
      Tags:
        type: "string"
      AppVersion:
        type: "string"
      Deprecated:
        type: "boolean"
      TillerVersion:
        type: "string"
      Annotations:
        $ref: "#/definitions/chart.Map"
      KubeVersion:
        type: "string"
  chart.Template:
    type: "object"
    properties:
      Name:
        type: "string"
      Data:
        type: "string"
        format: "byte"
  chart.Maintainer:
    type: "object"
    properties:
      Name:
        type: "string"
      Email:
        type: "string"
      Url:
        type: "string"
  chart.Config:
    type: "object"
    properties:
      Raw:
        type: "string"
      Values:
        $ref: "#/definitions/chart.Map"
  chart.Map:
    type: "array"
    items:
      $ref: "#/definitions/chart.Value"
  chart.Value:
    type: "object"
    properties:
      Key:
        type: "string"
      Value:
        type: "string"
  protobuf.Any:
    type: "object"
    properties:
      Type_Url:
        type: "string"
      Value:
        type: "string"
        format: "byte"
  APIResponse:
    type: "object"
    properties:
      code:
        type: "string"
      message:
        type: "string"
    required:
      - code
      - message
responses:
...
  NotFound:
    description: "The specified resource was not found"
    schema:
      $ref: "#/definitions/APIResponse"
  Unauthorized:
    description: "Unauthorized"
    schema:
      $ref: "#/definitions/APIResponse"
  BadRequest:
    description: "Bad Request"
    schema:
      $ref: "#/definitions/APIResponse"
...
  InternalServerError:
    description: "Internal Server Error"
    schema:
      $ref: "#/definitions/APIResponse"